Added many features to C Code Generator.
Huh so I have forgotten to commit for a long time even though we have made a bunch of
changes!!!

This is a good place to commit though. We added support for:
Timers: We can read from time, wait on timers.
Alts: We can Use Alt block to make choices between statements.